remove sync version of WaitForXXX
