+++ .github/workflows/verification.yml
@@ -61,10 +61,8 @@ jobs:
 
       - name: 🧪 Run unit tests (async)
         run: |
-          dotnet test --filter Category!=sync -c release --blame-hang-timeout 15s --blame-hang-dump-type full --blame-crash-dump-type full
-      - name: 🧪 Run unit tests (sync)
-        run: |
-          dotnet test --filter Category!=async -c release --blame-hang-timeout 15s --blame-hang-dump-type full --blame-crash-dump-type full
+          dotnet test --filter -c release --blame-hang-timeout 15s --blame-hang-dump-type full --blame-crash-dump-type full
+
       - name: 📛 Upload hang- and crash-dumps on test failure
         if: failure()
         uses: actions/upload-artifact@v3
