+++ tests/Directory.Build.props
@@ -8,7 +8,7 @@
 
 	<PropertyGroup Label="Compile settings" Condition="$(MSBuildProjectName) != 'bunit.testassets'">
 		<Nullable>annotations</Nullable>
-		<IsPackable>false</IsPackable>		
+		<IsPackable>false</IsPackable>
 		<SonarQubeTestProject>true</SonarQubeTestProject>
 		<IsTestProject>true</IsTestProject>
 	</PropertyGroup>
@@ -16,6 +16,7 @@
 	<ItemGroup Condition="$(MSBuildProjectName) != 'bunit.testassets'">
 		<PackageReference Include="AutoFixture.AutoNSubstitute" Version="4.17.0" />
 		<PackageReference Include="AutoFixture.Xunit2" Version="4.17.0" />
+		<PackageReference Include="NSubstitute" Version="4.3.0" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
 		<PackageReference Include="Moq" Version="4.18.2" />
 		<PackageReference Include="Shouldly" Version="4.1.0" />
